requirements.txt
pyyaml
pytest
prometheus-client
os